close add-on window and return when opening a repo
Firefox doesn't automatically close the add-on window when creating a new tab, so we need to do it manually. Also, we need to return from the `onFormSubmit` function when opening a tab, as opening the tab will not end its execution, and we might see wrong error messages in the add-on window.