Don't open new tab is the current tab is empty
Sometimes the user opens a new tab to go to some repo, and then recall she can use Eyfo to open that repo, so she opens Eyfo and selects the repo she wants. With this change, it will open the repo in the current blank tab, instead of opening a new tab.
(Tested only on Firefox)